optimize?

+++ gbg_program/_Save_Recall.ino
@@ -647,9 +647,9 @@ void recallSettings()
     while (true) { // flash SOS forever
       Serial.println(F("{\"error\": \"eeprom failure\"}"));
 
-      for (int j = 0; j < 3; j++) { // S O S
-        int t = (j == 1) ? 600 : 200;
-        for (int i = 0; i < 3; i++) {
+      for (byte j = 0; j < 3; j++) { // S O S
+        int16_t t = (j == 1) ? 600 : 200;
+        for (byte i = 0; i < 3; i++) {
           LED_ON; delay(t); LED_OFF; delay(200); 
 #if defined(IS_PICO)
           rp2040.wdt_reset();
